refactored track preview
